requirements.txt
ansible>=2.0,<2.2
jinja2>=2.7
colorlog>=2.6.1
gitpython>=2.0.8
